app/requirements.txt
anyio==3.6.1
appnope==0.1.3
appscript==1.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
Brotli==1.0.9
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
dash==2.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
fastapi==0.78.0
fastjsonschema==2.15.3
Flask==2.1.2
Flask-Compress==1.12
greenlet==1.1.2
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.5.1
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
lxml==4.9.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.11
numexpr==2.8.1
numpy==1.22.3
numpyencoder==0.3.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
plotly==5.8.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.0
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.9.1
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==23.0.0
requests==2.27.1
rfc3986==1.5.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.36
stack-data==0.2.0
starlette==0.19.1
tables==3.7.0
tenacity==8.0.1
terminado==0.15.0
tinycss2==1.1.1
tornado==6.1
traitlets==5.2.1.post0
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
uvicorn==0.17.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
websockets==10.3
Werkzeug==2.1.2
widgetsnbextension==3.6.0
xlwings==0.27.10
zipp==3.8.0
